removed old files and added comments
