+++ frontend/src/main.tsx
@@ -2,14 +2,12 @@ import { StrictMode } from 'react'
 import { createRoot } from 'react-dom/client'
 import { createBrowserRouter, RouterProvider } from "react-router-dom";
 import './index.css'
-import Profile from './pages/Homepage.tsx'
 import Login from './pages/Login.tsx';
-import { UsernameProvider } from './contexts/UsernameContextProvider.tsx';
-import Post from './pages/PostPage.tsx';
 import { DataProvider } from './contexts/DataContextProvider.tsx';
 import UserProfile from './pages/UserProfilePage.tsx';
-import CreatePost from './pages/CreatePostPopup.tsx';
+import CreatePost from './pages/CreatePostPage.tsx';
 
+// Defines different URL paths for website
 const router = createBrowserRouter([
   {
     path: "/",
@@ -23,19 +21,12 @@ const router = createBrowserRouter([
     path: "create",
     element: <CreatePost />,
   },
-  {
-    path: "post/:id",
-    element: <Post/>,
-  }
 ]);
- 
 
 createRoot(document.getElementById('root')!).render(
   <StrictMode>
     <DataProvider>
-      <UsernameProvider>
         <RouterProvider router={router} />
-      </UsernameProvider>
     </DataProvider>
   </StrictMode>,
 )
